+++ weave-js/src/components/Panel2/PanelPlot/types.ts
@@ -3,6 +3,13 @@ import * as TableType from '../PanelTable/tableType';
 import {DimensionLike} from './plotState';
 import {AnyPlotConfig, PlotConfig} from './versions';
 
+export type AxisName = 'x' | 'y';
+
+export type VegaTimeUnit = 'yearweek' | 'yearmonth';
+
+export const BRUSH_MODES = ['zoom' as const, 'select' as const];
+export type BrushMode = (typeof BRUSH_MODES)[number];
+
 export const inputType = TableType.GeneralTableLikeType;
 export type PanelPlotProps = Panel2.PanelProps<typeof inputType, AnyPlotConfig>;
